requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.5
python-dotenv==1.0.0
openai==0.27.2
